Time: 52 ms (17.13%), Space: 125.2 MB (65.71%) - LeetHub

+++ 2300-successful-pairs-of-spells-and-potions/2300-successful-pairs-of-spells-and-potions.cpp
@@ -7,7 +7,7 @@ public:
         }
         return s;
     }
-    int help(vector<int>& p,long long x,int target){
+    int help(vector<int>& p,long long x,long long target){
         int n = p.size();
         int low = 0, high = n - 1, ans = n;
